k40-usb/requirements.txt
rx=3.0.0a3
pyusb==1.0.2